src/main/java/ejercicioP/service/MovieService.java
package ejercicioP.service;

import ejercicioP.data.MovieRepository;
import ejercicioP.model.Genre;
import ejercicioP.model.Movie;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.stream.Collectors;

@Service
public class MovieService {

    private MovieRepository movieRepository;

    public MovieService(MovieRepository movieRepository) {
        this.movieRepository = movieRepository;
    }

    public List<Movie> findMoviesByGenre(Genre genre) {

        return movieRepository.findAll().stream()
                .filter(movie -> movie.getGenre()==genre).collect(Collectors.toList());
    }

    public List<Movie> findMoviesByLength(int length) {

        return movieRepository.findAll().stream()
                .filter(movie -> movie.getMinutes()<=length).collect(Collectors.toList());
    }
}
